Restrict NER scanning for Python source code to comments and strings
If a filename smells like Python, then extract the string and comment tokens
from the source and feed only those parts to the named entity recognizer,
ignoring things like Python function names.

This should reduce the number of false positives for PII in Python source code.

The `CheckFileBase` class is extended by the other checks, not just the NER
checker, but only the `CheckForNER` class needs to switch strategy for a file
depending on the filename, so I have changed the relevant method signature to
pass the filename (as well as the file content).