Restore support for Python 3.7 and 3.8
Some projects use older Python versions, and devs don't have Python 3.9 setup.
This change removes the explicit requirement for Python 3.9 in the hook config.

The code was using pathlib's `is_relative_to(..)` which was added in 3.9, so
I've added the equivalent compatible code for older versions.

Tests were broken on Py 3.7 cos of `\s` in a string, so that was fixed by
marking the string as a raw string. The tests also took a super long time to
setup using tox on Py 3.7 because the dependencies were installed twice.